fixing format for test
