+++ networkx/algorithms/approximation/maximum_weight_cycle_packing_approximation_algorithm.py
@@ -1,12 +1,7 @@
 import doctest
-import logging
 
 import networkx as nx
-
-# logging.basicConfig(filemode="Logs.log", level=logging.INFO)
-# log = logging.getLogger()
-
-EPS = 1e-6
+from networkx.algorithms.simple_cycles_le_k import simple_cycles_le_k
 
 """REUT HADAD & TAL SOMECH"""
 
@@ -15,7 +10,7 @@ This is an implementation for two different algorithms described on "MAXIMUM WEI
 WITH APPLICATION TO KIDNEY EXCHANGE PROGRAMS" article.
 The article points on two algorithms that solves kidney exchange problems, which can be modelled as cycle packing
 problems in a directed graph, involving cycles of length 2, 3, or even longer.
-In the article we focus on the maximal exchange of circles of size 2 and 3 vertices, we demonstrate an approximation
+In the article we focus on the maximal exchange of cycles of size 2 and 3 vertices, we demonstrate an approximation
 algorithm and an exact algorithm for this problem.
 """
 """article title: MAXIMUM WEIGHT CYCLE PACKING IN DIRECTED GRAPHS,WITH APPLICATION TO KIDNEY EXCHANGE PROGRAMS
@@ -23,16 +18,19 @@ authors:Biro, P. and Manlove, D.F. and Rizzi, R.
 year:(2009)
 link:http://eprints.gla.ac.uk/25732/
 """
+EPS = 1e-6
+"""The EPS parameter is used in the maximum weight independent set algorithm in which the epsilon stands for
+    algorithm accuracy."""
 
 
-def Maximum_weight_cycle_packing_approximation_algorithm(graph: nx.DiGraph, k: int):
+def maximum_weight_cycle_packing_approximation_algorithm(graph: nx.DiGraph, k: int):
     """
     "
     A directed weighted graph is a graph in which every edge is one sided and weighted
     for example an edge from node 1->2 with a weight of 5,an k-way exchange
-    is a circle within a graph containing at most k nodes.
-    max weighted exchange is a circle with the most weighted edges from every node in the circle
-
+    is a cycle within a graph containing at most k nodes.
+    max weighted exchange is a set of cycles, such that no node appears in more than one cycle,
+    and the sum of weights on the edges in all cycles is maximum.
     Parameters
     -----------
     G : NetworkX DiGraph
@@ -41,77 +39,79 @@ def Maximum_weight_cycle_packing_approximation_algorithm(graph: nx.DiGraph, k: i
     Returns
     -----------
     Lst: list of lists
-        Each list in lst contaning the nodes which make up the circle with the highest weights sum
+        Each list in lst contaning the nodes which make up a single cycle.
+        The sum of weights on the edges in all cycles together is an approximation of the maximum.
     Examples
     -----------
     >>> Digraph=nx.DiGraph()
     >>> Digraph.add_nodes_from([1,2,3,4,5,6,7,8])
     >>> Digraph.add_weighted_edges_from([(1,8,2),(8,1,4),(2,1,5),(1,3,4),(3,8,2),(8,2,3),(8,5,4),(5,7,3),(7,6,2),(6,5,4)])
-    >>> print(Maximum_weight_cycle_packing_approximation_algorithm(Digraph, 3))
+    >>> print(maximum_weight_cycle_packing_approximation_algorithm(Digraph, 3))
     [(5, 7, 6), (8, 2, 1)]
     >>> Digraph =nx.DiGraph()
     >>> Digraph.add_nodes_from([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20])
     >>> Digraph.add_weighted_edges_from([(1, 6, 11), (6, 1, 10), (1, 5, 3), (5, 1, 2), (8, 9, 11), (9, 8, 20), (3, 2, 6), (2, 6, 5), (6, 3, 8),(5, 7, 6), (7, 4, 11), (4, 5, 5), (10, 16, 1), (16, 11, 10), (11, 15, 3), (15, 11, 2), (18, 19, 11),(19, 18, 20), (13, 12, 6), (12, 16, 5), (16, 13, 8)])
-    >>> print(Maximum_weight_cycle_packing_approximation_algorithm(Digraph, 2))
+    >>> print(maximum_weight_cycle_packing_approximation_algorithm(Digraph, 2))
     [(1, 6), (8, 9), (11, 15), (18, 19)]
     >>> Digraph=nx.DiGraph()
     >>> Digraph.add_nodes_from([1,2,3,4,5,6,7,8])
     >>> Digraph.add_weighted_edges_from([(8,1,4),(2,1,5),(1,3,4),(3,8,2),(8,2,3),(8,5,4),(5,7,3),(7,6,2),(6,5,4)])
-    >>> print(Maximum_weight_cycle_packing_approximation_algorithm(Digraph, 2))
+    >>> print(maximum_weight_cycle_packing_approximation_algorithm(Digraph, 2))
     []
 
 
     Notes
     -----------
-    Algorithm - the algorithm finds maximum weight k-way exchanges using reduction from graph to hyper-graph by
+    Algorithm - the algorithm finds an approximate maximum weight k-way exchanges using reduction from graph to hyper-graph by
     the algorithm in the published article APX-complete algorithm for kidney exchange programs"
 
-    Refrences
+    References
     ----------
     Algorithm 1 - 'MAXIMUM WEIGHT CYCLE PACKING IN DIRECTED GRAPHS, WITH APPLICATION TO KIDNEY EXCHANGE PROGRAMS' by Biro, P. and Manlove, D.F. and Rizzi, R. http://eprints.gla.ac.uk/25732/
-    :param k:  num of circles size
+    :param k:   maximum cycle size.
     :param graph: graph that simulates donors-Patients for kidney exchange programs.
 
     """
-    logging.info("Finds cycles of size k in graph")
-    cycles = find_cycles_of_size_k(graph, k)
+    # logging.info(f"Find all cycles of size at most {k} in the graph")
+
+    sc = simple_cycles_le_k(graph, k)
+    cycles = []
+    for cycle in sc:
+        cycles.append(tuple(cycle))
 
     if len(cycles) == 0:
         return []
 
-    logging.info(
-        "create hypergraph according to hypergraph's format of hypernetx library - "
-        "https://pnnl.github.io/HyperNetX/build/classes/classes.html"
-        "{i,Entity{i,([nodesOfSingleHyperEdge]),"
-        "'weight'}}"
-    )
+    #     "create hypergraph according to hypergraph's format of hypernetx library - "
+    #     "https://pnnl.github.io/HyperNetX/build/classes/classes.html"
+    #     "{i,Entity{i,([nodesOfSingleHyperEdge]),"
+    #     "'weight'}}"
     buildHype = {}
     weights = {}
 
     for i, cycle in enumerate(cycles):
         if len(cycle) < 3:
             weight = (
-                graph.get_edge_data(cycle[0], cycle[1])["weight"]
-                + graph.get_edge_data(cycle[1], cycle[0])["weight"]
+                    graph.get_edge_data(cycle[0], cycle[1])["weight"]
+                    + graph.get_edge_data(cycle[1], cycle[0])["weight"]
             )
         else:
             weight = (
-                graph.get_edge_data(cycle[0], cycle[1])["weight"]
-                + graph.get_edge_data(cycle[1], cycle[2])["weight"]
-                + graph.get_edge_data(cycle[2], cycle[0])["weight"]
+                    graph.get_edge_data(cycle[0], cycle[1])["weight"]
+                    + graph.get_edge_data(cycle[1], cycle[2])["weight"]
+                    + graph.get_edge_data(cycle[2], cycle[0])["weight"]
             )
         buildHype[cycle] = weight
         weights[i] = weight
 
-    logging.info("create a new graph for maximum weight independent set algorithm")
+    # "create a new graph for maximum weight independent set algorithm"
     graphL = nx.Graph()
 
     for k, v in buildHype.items():
         graphL.add_node(k, weight=v)
 
-    logging.info(
-        "create appearance of nodes that related to each node in the directed graph"
-    )
+    # "create appearance of nodes that related to each node in the directed graph"
+
     nodes = graph.nodes
     temp = {}
     for node in nodes:
@@ -121,7 +121,7 @@ def Maximum_weight_cycle_packing_approximation_algorithm(graph: nx.DiGraph, k: i
                 temp2.append(edge)
         temp[node] = temp2
 
-    logging.info("add edges to undirected graph")
+    #   "add edges to undirected graph"
     for k, v in temp.items():
         if len(v) > 1:
             for i in range(0, len(v)):
@@ -129,104 +129,12 @@ def Maximum_weight_cycle_packing_approximation_algorithm(graph: nx.DiGraph, k: i
                     if i is not len(v):
                         graphL.add_edge(v[i], v[j])
 
-    logging.info("nodes and weights")
+    #   "nodes and weights"
     pi = dict(zip(graphL.nodes(), (weights[i] for i in weights)))
 
-    logging.info("find maximum weight independent set")
+    #   "find maximum weight independent set"
     mwis_set, mwis_weight = MWIS(graphL, pi, 0)
     return mwis_set
-
-
-def simple_cycles(graph, k):
-    """
-    >>> Digraph=nx.DiGraph()
-    >>> Digraph.add_nodes_from([1,2,3,5,6,7,8])
-    >>> Digraph.add_weighted_edges_from([(1,8,2),(8,1,4),(2,1,5),(1,3,4),(3,8,2),(8,2,3),(8,5,4),(5,7,3),(7,6,2),(6,5,4)])
-    >>> Ys=simple_cycles(Digraph,3)
-    >>> a=[y for y in Ys]
-    >>> print(a)
-    [[8, 2, 1], [8, 1, 3], [8, 1], [5, 7, 6]]
-    >>> Digraph =nx.DiGraph()
-    >>> Digraph.add_nodes_from([1,2,3,4])
-    >>> Digraph.add_weighted_edges_from([(2,1,3),(1,3,1),(3,2,2),(3,4,5),(4,3,9)])
-    >>> Ys=simple_cycles(Digraph,3)
-    >>> a=[y for y in Ys]
-    >>> print(a)
-    [[1, 3, 2], [3, 4]]
-    >>> graphEX3 = nx.DiGraph()
-    >>> graphEX3.add_nodes_from([10,11,12,13,14,15,16])
-    >>> Digraph.add_weighted_edges_from([(10,11,10),(11,12,5),(12,13,6),(13,10,4),(11,14,2),(14,16,3),(16,15,8),(15,14,6)])
-    >>> Ys=simple_cycles(Digraph,3)
-    >>> a=[y for y in Ys]
-    >>> print(a)
-    [[16, 15, 14], [1, 3, 2], [3, 4]]
-    """
-    subG = type(graph)(graph.edges())
-    ccs = list(nx.strongly_connected_components(subG))
-    while ccs:
-        scc = ccs.pop()
-        startnode = scc.pop()
-        path = [startnode]
-        blocked = set()
-        blocked.add(startnode)
-        stack = [(startnode, list(subG[startnode]))]
-
-        while stack:
-            thisnode, nbrs = stack[-1]
-
-            if nbrs and len(path) <= k:
-                nextnode = nbrs.pop()
-                if nextnode == startnode:
-                    yield path[:]
-                elif nextnode not in blocked:
-                    path.append(nextnode)
-                    stack.append((nextnode, list(subG[nextnode])))
-                    blocked.add(nextnode)
-                    continue
-            if not nbrs or len(path) >= k:
-                blocked.remove(thisnode)
-                stack.pop()
-                path.pop()
-        subG.remove_node(startnode)
-        H = subG.subgraph(scc)
-        ccs.extend(nx.strongly_connected_components(H))
-
-
-def find_cycles_of_size_k(graph, k):
-    """
-    finds relevent cycles according to chosen K
-    Parameters
-    -----------
-    :param graph : NetworkX DiGraph
-        Directed graph with weights
-    +    :param k : size of cycles (according to article - max size is 3 and min size is 2)
-    Returns
-    -----------
-    cycles: cycles on graph according to chosen K
-    Examples
-    -----------
-    >>> Digraph=nx.DiGraph()
-    >>> Digraph.add_nodes_from([1,2,3,4,5,6,7,8])
-    >>> Digraph.add_weighted_edges_from([(1,8,2),(8,1,4),(2,1,5),(1,3,4),(3,8,2),(8,2,3),(8,5,4),(5,7,3),(7,6,2),(6,5,4)])
-    >>> print(find_cycles_of_size_k(Digraph, 3))
-    [(8, 2, 1), (8, 1, 3), (8, 1), (5, 7, 6)]
-    >>> Digraph =nx.DiGraph()
-    >>> Digraph.add_nodes_from([1,2,3,4])
-    >>> Digraph.add_weighted_edges_from([(2,1,3),(1,3,1),(3,2,2),(3,4,5),(4,3,9)])
-    >>> print(find_cycles_of_size_k(Digraph, 2))
-    [(3, 4)]
-    >>> Digraph=nx.DiGraph()
-    >>> Digraph.add_nodes_from([1,2,3,4,5,6,7,8])
-    >>> Digraph.add_weighted_edges_from([(8,1,4),(2,1,5),(1,3,4),(3,8,2),(8,2,3),(8,5,4),(5,7,3),(7,6,2),(6,5,4)])
-    >>> print(find_cycles_of_size_k(Digraph, 2))
-    []
-    """
-    sc = simple_cycles(graph, k)
-    cycles = []
-    for cycle in sc:
-        if 1 < len(cycle) <= k:
-            cycles.append(tuple(cycle))
-    return cycles
 
 
 """recursive heuristic for maximum weighted independent sets
