+++ networkx/algorithms/approximation/tests/test_maximum_weight_cycle_packing_approximation.py
@@ -1,7 +1,7 @@
 import networkx as nx
 import random
-from networkx.algorithms.maximum_weight_cycle_packing import (
-    maximum_weight_cycle_packing,
+from networkx.algorithms.approximation.maximum_weight_cycle_packing_approximation_algorithm import (
+    maximum_weight_cycle_packing_approximation_algorithm,
 )
 
 
@@ -24,16 +24,21 @@ def test_1():
             (5, 6, 12),
         ]
     )
-    assert maximum_weight_cycle_packing(graphEX3, 3) == [[1, 6, 4], [2, 3]]
-    assert maximum_weight_cycle_packing(graphEX3, 2) == [[1, 6, 4]]
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3) == [
+        (1, 6, 4),
+        (2, 3),
+    ]
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 2) == [
+        (1, 6),
+        (2, 3),
+    ]
 
 
 def test_2():
     graphEX3 = nx.DiGraph()
     graphEX3.add_nodes_from([1, 2, 3])
     graphEX3.add_weighted_edges_from([(1, 2, 1), (2, 1, 5), (2, 3, 2), (3, 1, 2)])
-    assert maximum_weight_cycle_packing(graphEX3, 3) == [[1, 2, 3]]
-    assert maximum_weight_cycle_packing(graphEX3, 2) == [[2, 1]]
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 2) == [(1, 2)]
 
 
 def test_3():
@@ -55,7 +60,11 @@ def test_3():
             (4, 5, 5),
         ]
     )
-    assert maximum_weight_cycle_packing(graphEX3, 3) == [[2, 6, 3], [8, 9], [4, 5, 7]]
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3) == [
+        (1, 6),
+        (4, 5, 7),
+        (8, 9),
+    ]
 
 
 def test_4():
@@ -88,21 +97,19 @@ def test_4():
             (16, 13, 8),
         ]
     )
-    assert maximum_weight_cycle_packing(graphEX3, 3) == [
-        [16, 13, 12],
-        [4, 5, 7],
-        [2, 6, 3],
-        [11, 15],
-        [8, 9],
-        [18, 19],
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3) == [
+        (1, 6),
+        (4, 5, 7),
+        (8, 9),
+        (11, 15),
+        (16, 13, 12),
+        (18, 19),
     ]
-    assert maximum_weight_cycle_packing(graphEX3, 3) == [
-        [16, 13, 12],
-        [4, 5, 7],
-        [2, 6, 3],
-        [11, 15],
-        [8, 9],
-        [18, 19],
+    assert maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 2) == [
+        (1, 6),
+        (8, 9),
+        (11, 15),
+        (18, 19),
     ]
 
 
@@ -110,7 +117,7 @@ def test_5():
     graphEX3 = nx.fast_gnp_random_graph(20, 0.15, 42, True)
     for (u, v, w) in graphEX3.edges(data=True):
         w['weight'] = random.randint(0, 10)
-    res = maximum_weight_cycle_packing(graphEX3, 3)
+    res = maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3)
     nodes_seen = []
     flag = True
     for cyc in res:
@@ -126,7 +133,7 @@ def test_6():
     graphEX3 = nx.fast_gnp_random_graph(20, 0.15, 42, True)
     for (u, v, w) in graphEX3.edges(data=True):
         w['weight'] = random.randint(0, 10)
-    res = maximum_weight_cycle_packing(graphEX3, 3)
+    res = maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3)
     nodes_seen = []
     flag = True
     for cyc in res:
@@ -144,7 +151,7 @@ def test_7():
     sc = simple_cycles_le_k(graphEX3, 3)
     try:
         cy = next(sc)
-        res = maximum_weight_cycle_packing(graphEX3, 3)
+        res = maximum_weight_cycle_packing_approximation_algorithm(graphEX3, 3)
         if cy is not None:
             assert len(res) > 1
     except:
